% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Master_Summary.R
\name{Master_Summary}
\alias{Master_Summary}
\title{Table Recapping all Relevant Location and Movement Metrics}
\usage{
Master_Summary(data)
}
\arguments{
\item{data}{Trackman data with standard naming}
}
\value{
A table with all metrics averaged and color-coded strike percentage calculations
}
\description{
Table Recapping all Relevant Location and Movement Metrics
}
\examples{
library(dplyr)
library(ggplot2)
library(knitr)
library(kableExtra)
#mini sample trackman dataset
demo_trackman_data <- data.frame(
  TaggedPitchType = c("Fastball", "Curveball"),
  RelSpeed = c(88.23, 80.35),
  InducedVertBreak = c(15.25, -6.44),
  HorzBreak = c(12.87, -3.45),
  PlateLocHeight = c(1.73, 2.14),
  PlateLocSide = c(0.84, 0.92))
 Master_Summary(demo_trackman_data)
}
