% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pitch_Plot.R
\name{Pitch_Plot}
\alias{Pitch_Plot}
\title{Generating Pitch Movement Plot and Summary Table}
\usage{
Pitch_Plot(data)
}
\arguments{
\item{data}{Trackman data with standard naming}
}
\value{
A summary table with session averages and advanced stats like VAA, as well as a pitch movement plot
}
\description{
Generating Pitch Movement Plot and Summary Table
}
\examples{
library(dplyr)
library(ggplot2)
library(knitr)
#mini sample trackman dataset
demo_trackman_data <- data.frame(
  TaggedPitchType = c("Fastball", "Curveball"),
  RelSpeed = c(88.23, 80.35),
  InducedVertBreak = c(15.25, -6.44),
  HorzBreak = c(12.87, -3.45))

 Pitch_Plot(demo_trackman_data)
}
