% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plate_Viz.R
\name{Plate_Viz}
\alias{Plate_Viz}
\title{Pitch Location Plot}
\usage{
Plate_Viz(data)
}
\arguments{
\item{data}{Trackman data with standard naming}
}
\value{
a plot displaying pitch location within the strikezone to see command and location
}
\description{
Pitch Location Plot
}
\examples{
#mini sample trackman dataset
library(dplyr)
library(ggplot2)
library(knitr)
demo_trackman_data <- data.frame(
  TaggedPitchType = c("Fastball", "Curveball"),
  RelSpeed = c(88.23, 80.35),
  PlateLocHeight = c(1.73, 2.14),
  PlateLocSide = c(0.84, 0.92))

 Plate_Viz(demo_trackman_data)
}
