% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackman_data.R
\docType{data}
\name{trackman_data}
\alias{trackman_data}
\title{A sample Trackman pitching dataset to highlight the funcitons in BSBReport}
\format{
\subsection{\code{trackman_data}}{

A data frame with 25 rows and 8 columns:
\describe{
\item{PitcherId}{Integer value denoting unique player id}
\item{TaggedPitchType}{Character denoting pitch type}
\item{PitchCount}{integer of pitch count}
\item{InducedVertBreak}{integer representing the IVB for each pitch}
\item{HorzBreak}{integer representing the HB for each pitch}
\item{PlateLocHeight}{integer of measured pitch height at front of home plate}
\item{PlateLocSide}{integer of measured pitch horizontal location at front of home plate}
\item{RelSpeed}{integer of measured pitch speed at release in mph}
}
}
}
\usage{
trackman_data
}
\description{
A sample Trackman pitching dataset to highlight the funcitons in BSBReport
}
\keyword{datasets}
